Cleanup startup - moved config into middleware extension methods
